profile loads on login but has no access to user info yet
