+++ src/components/Layout/Footer.jsx
@@ -1,7 +1,7 @@
 import { Button, Space, Typography, Image } from 'antd';
 import React from 'react';
 import Logo from '../../styles/Images/WhiteLogo.png';
-import '../../styles/LessRenders/RenderFooter.less';
+import '../../styles/LessRenders/Layout/Footer.less';
 
 const { Text } = Typography;
