+++ src/components/pages/Landing/RenderLandingPage.jsx
@@ -3,7 +3,7 @@ import GrantRatesByOfficeImg from '../../../styles/Images/bar-graph-no-text.png'
 import GrantRatesByNationalityImg from '../../../styles/Images/pie-chart-no-text.png';
 import GrantRatesOverTimeImg from '../../../styles/Images/line-graph-no-text.png';
 import HrfPhoto from '../../../styles/Images/paper-stack.jpg';
-import '../../../styles/LessRenders/RenderLandingPage.less';
+import '../../../styles/LessRenders/Pages/LandingPage.less';
 import { Button } from 'antd';
 import { useHistory } from 'react-router-dom';
 // for the purposes of testing PageNav
