--- src/components/pages/Protected/LoadingPage.js
@@ -1,15 +1,0 @@
-// Reference: https://developer.auth0.com/resources/guides/spa/react/basic-authentication/v17-javascript-react-router-5
-
-import React from 'react';
-// import { NavBar } from "../components/navigation/desktop/nav-bar";
-// import { MobileNavBar } from "../components/navigation/mobile/mobile-nav-bar";
-
-export const LoadingPage = () => {
-  return (
-    <div className="page-layout">
-      {/* <NavBar />
-      <MobileNavBar /> */}
-      <div className="page-layout__content" />
-    </div>
-  );
-};
