+++ src/styles/LessRenders/Layout/Header.less
@@ -22,7 +22,7 @@ header.ant-layout-header {
       .header-button {
         background: none!important;
         border: none;
-        padding: 0 25px!important;
+        margin: 0 25px!important;
         color: #E2F0F7;
         cursor: pointer;
       }
