+++ src/styles/LessRenders/Pages/ProfilePage.less
@@ -1,0 +1,19 @@
+div#profile-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  padding: 5%;
+
+  .button-container {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+
+    button {
+      margin: 2%;
+    }
+  }
+
+}
